% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_new_type.R
\name{as_assertion_factory}
\alias{as_assertion_factory}
\title{Build a new type}
\usage{
as_assertion_factory(f)
}
\arguments{
\item{f}{a function}
}
\value{
a function with class \code{assertion_factory}
}
\description{
Build a new type
}
