% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_assert_types.R
\name{assert_types}
\alias{assert_types}
\title{Assert Types}
\usage{
assert_types()
}
\description{
This function shouldn't be called by the user, it will be inserted in the
body of a function to check the arguments and convert them to active bindings
if relevant.
}
