% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{assign}
\alias{assign}
\alias{<-}
\title{Typed assign}
\usage{
(x) <- value
}
\description{
Drop-in replacement for standard assignment to parse input
}
\examples{
`at<-` <- function(a, b, c){
a[b[[1]], b[[2]]] <- c
a
}
mat <- matrix(0, 3, 3)
mat[2, 2] <- 7
at(mat, c(1,2)) <- 5
}
