% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_declare.R
\name{check_output}
\alias{check_output}
\alias{check_arg}
\title{Check Argument Types and Return Type}
\usage{
check_output(.output, .assertion, ...)

check_arg(.arg, .assertion, ..., .bind = FALSE)
}
\arguments{
\item{.output}{function output}

\item{.assertion}{an assertion}

\item{...}{additional arguments passed to assertion}

\item{.arg}{function argument}

\item{.bind}{whether to actively bind the argument so it cannot be modified
unless it satisfies the assertion}
}
\value{
\code{.output}if it satisfies the assertion, fails otherwise.

returns \code{NULL} invisibly, called for side effects.
}
\description{
These functions are not designed to be used directly, we advise to use the
syntaxes described in \code{?declare} instead. \code{check_arg} checks that arguments
satisfy an assertion, and if relevant make them into active bindings to make sure they
always satisy it. \code{check_output} checks that the value, presumably a return
value, satisfies an assertion,
}
