% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_ts.R
\name{compile}
\alias{compile}
\title{Compile Typescript definitions}
\usage{
compile(x, file = sprintf("\%s.d.ts", deparse(substitute(x))))
}
\arguments{
\item{x}{Typed function to compile}

\item{file}{File to write to}
}
\value{
NULL - writes a file
}
\description{
Compile Typescript definitions
}
