% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_const.R, R/main3.R
\name{const}
\alias{const}
\title{Define Constants}
\usage{
const(x)

const(x)
}
\arguments{
\item{x}{value}
}
\description{
\code{const()} is meant to be used using the syntax \code{? x <- const(3)} or \code{? x <- const(3)}.
The variable then cannot be modified anymore.

\code{const()} is meant to be used using the syntax \code{? x <- const(3)} or \code{? x <- const(3)}.
The variable then cannot be modified anymore.
}
