% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_new_type.R
\name{get_assertion}
\alias{get_assertion}
\title{Fetch assertion from active binding function}
\usage{
get_assertion(x)
}
\arguments{
\item{x}{obect}
}
\description{
Fetch assertion from active binding function
}
