% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_new_type.R
\name{get_assertion_fun}
\alias{get_assertion_fun}
\title{Fetch assertion function from active binding function}
\usage{
get_assertion_fun(x)
}
\arguments{
\item{x}{obect}
}
\description{
Fetch assertion function from active binding function
}
