% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_question_mark.R, R/main3.R
\name{?}
\alias{?}
\title{Static typing}
\usage{
`?`(lhs, rhs)

`?`(lhs, rhs)
}
\arguments{
\item{lhs}{lhs}

\item{rhs}{rhs}
}
\description{
Use \verb{?} to set a function's return type, argument types, or variable types
in the body of the function.

Use \verb{?} to set a function's return type, argument types, or variable types
in the body of the function.
}
\examples{
numeric ? add2 <- function (x= ?numeric) {
  numeric ? y <- 2
  res <- x + y
  res
}
add2
add2(3)
numeric ? add2 <- function (x= ?numeric) {
  numeric ? y <- 2
  res <- x + y
  res
}
add2
add2(3)
}
