% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locked.R
\name{locked}
\alias{locked}
\alias{locked<-}
\alias{immutable<-}
\title{Lock variable binding}
\usage{
locked(x) <- value

immutable(x) <- value
}
\description{
Using `locked` (or its alias, `immutable`) together with `typed` package
implementation of assignment operator allows you to lock variable binding in simplier, more
straighforward way (in one function call, without additional `lockBinding` function calls).
}
\examples{
x <- 10
x <- 2.5
x
locked(q) <- 20
q <- 10
123 -> locked(y)
y <- 25
}
