% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_new_type.R
\name{process_assertion_factory_dots}
\alias{process_assertion_factory_dots}
\title{Process assertion factory dots}
\usage{
process_assertion_factory_dots(...)
}
\arguments{
\item{...}{dots}
}
\description{
This needs to be exported, but shouldn't be called by the user
}
