% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_new_type.R
\name{process_assertions}
\alias{process_assertions}
\title{Process assertions}
\usage{
process_assertions(...)
}
\arguments{
\item{...}{additional assertions}
}
\description{
This needs to be exported, but shouldn't be called by the user
}
