% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_new_type.R
\name{process_type_checker_dots}
\alias{process_type_checker_dots}
\title{Process type checker dots}
\usage{
process_type_checker_dots(...)
}
\arguments{
\item{...}{dots}
}
\description{
This needs to be exported, but shouldn't be called by the user
}
