% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_question_mark.R, R/05_new_type.R
\name{?}
\alias{?}
\alias{declare}
\title{Static typing}
\usage{
`?`(lhs, rhs)

declare(x, assertion_fun, value)
}
\arguments{
\item{lhs}{lhs}

\item{rhs}{rhs}

\item{x}{variable name as a string}

\item{assertion_fun}{a function}

\item{value}{an optional value}
}
\description{
Use \verb{?} to set a function's return type, argument types, or variable types
in the body of the function.
}
\examples{
numeric ? function (x= ?numeric) {
  numeric ? y <- 2
  res <- x + y
  res
}
}
