% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed.R
\name{typed}
\alias{typed}
\title{Typed functions}
\usage{
typed(f, argtypes, value)
}
\arguments{
\item{f}{Function to type}

\item{argtypes}{List of types for each argument}

\item{value}{Return type}
}
\value{
Typed function
}
\description{
Typed functions
}
