% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{types}
\alias{types}
\alias{number}
\alias{string}
\alias{boolean}
\alias{void}
\alias{any_type}
\alias{litteral}
\alias{array_type}
\alias{union}
\alias{optional}
\alias{object}
\title{Type functions}
\usage{
number()

string()

boolean()

void()

any_type()

litteral(x)

array_type(type)

union(...)

optional(type)

object(...)
}
\arguments{
\item{x}{Litteral value}

\item{type}{Type to make optional}

\item{...}{Object components}
}
\description{
Use these functions to construct complex types.
}
\section{Functions}{
\itemize{
\item \code{number()}: Number type

\item \code{string()}: String type

\item \code{boolean()}: Boolean type

\item \code{void()}: Void type

\item \code{any_type()}: Any type

\item \code{litteral()}: Litteral type

\item \code{array_type()}: Array type

\item \code{union()}: Composite type

\item \code{optional()}: Optional types

\item \code{object()}: Object type

}}
