% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_use_typed.R
\name{use_typed}
\alias{use_typed}
\title{Use the 'typed' package}
\usage{
use_typed()
}
\value{
Returns \code{NULL} invisibly, called for side effects
}
\description{
This sets up your package so you can use type, by editing the DESCRIPTION file
and editing or creating 'R/your.pkg-package.R'
}
