% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_types.R
\name{validate_types}
\alias{validate_types}
\title{Check object x has valid type}
\usage{
validate_types(x, type = any())
}
\arguments{
\item{x}{Object to check}

\item{type}{Type to check against}
}
\value{
logical indicating if x has valid type
}
\description{
Check object x has valid type
}
