% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_question_mark.R, R/02_declare.R
\name{?}
\alias{?}
\alias{check_output}
\alias{check_arg}
\alias{declare}
\title{Static typing}
\usage{
`?`(lhs, rhs)

check_output(.output, .assertion, ...)

check_arg(.arg, .assertion, ..., .bind = FALSE)

declare(x, assertion, value, const = FALSE)
}
\arguments{
\item{lhs}{lhs}

\item{rhs}{rhs}

\item{.output}{function output}

\item{.assertion}{an assertion}

\item{...}{additional arguments passed to assertion}

\item{.arg}{function argument}

\item{.bind}{whether to actively bind the argument so it cannot be modified
unless it satisfies the assertion}

\item{x}{variable name as a string}

\item{assertion}{a function}

\item{value}{an optional value}

\item{const}{whether to declare \code{x} as a constant}
}
\description{
Use \verb{?} to set a function's return type, argument types, or variable types
in the body of the function.
}
\examples{
numeric ? function (x= ?numeric) {
  numeric ? y <- 2
  res <- x + y
  res
}
}
