% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed-package.R
\docType{package}
\name{typed-package}
\alias{typed}
\alias{typed-package}
\title{typed: Support Types for Variables, Arguments, and Return Values}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A type system for R. It supports setting variable types in a script or the body of a function, so variables can't be assigned illegal values. Moreover it supports setting argument and return types for functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moodymudskipper/typed}
  \item Report bugs at \url{https://github.com/moodymudskipper/typed/issues}
}

}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@gmail.com}

}
\keyword{internal}
